+++ sw.js
@@ -2,7 +2,7 @@
 //
 // Bump CACHE version whenever shipped files change so clients pick up the update.
 
-const CACHE = 'tripletown-v62';
+const CACHE = 'tripletown-v63';
 
 const SHELL = [
   '.',
@@ -21,6 +21,7 @@ const SHELL = [
   'js/render.js',
   'js/sprites.js',
   'js/gestures.js',
+  'js/villagers.js',
   'assets/icons/icon-192.png',
   'assets/icons/icon-512.png',
   'assets/icons/icon-180.png',